lawmadi-adk/requirements.txt
google-adk>=1.2.0
google-cloud-aiplatform[agent_engines,adk]>=1.112
requests>=2.31.0
